updated to production

+++ System/Trello.cs
@@ -1,5 +1,7 @@
 
 
+using System.Security.Cryptography.X509Certificates;
+
 public class Trello
 {
 
@@ -51,23 +53,78 @@ public class Trello
         var result = _boardService.AddBoard(board);
         return result;
     }
-    public Board RemoveBoard(Board board)
+    public Board RemoveBoard(string boardId)
     {
+        Board board=_boardService.GetBoard(boardId);
+        if(board == null){
+            _loggerService.Warning($"The board with  id {boardId} is not found.");
+            return null;
+        }
         Board result = _boardService.RemoveBoard(board);
         return result;
 
     }
-    public Board AddPeople(Board board, User user)
+    public Board AddMember(string boardId, string userId)
     {
+        Board board=_boardService.GetBoard(boardId);
+        User user=_userService.GetUser(userId);
+        if(board is null){
+            _loggerService.Error($"Board with Id {boardId} is not present.");
+            return null;
+        }
+        if(user is null){
+            _loggerService.Error($"user with id {userId} is not present. ");
+            return null;
+        }
         Board result = _boardService.AddMember(board, user);
         return result;
     }
 
-    public Board RemovePeople(Board board, User user)
+    public Board RemoveMember(string boardId, string userId)
     {
+        Board board=_boardService.GetBoard(boardId);
+        User user=_userService.GetUser(userId);
+        if(board is null){
+            _loggerService.Error($"Unable to find board with id {boardId}");
+            return null;
+        }
+        if(user is null){
+            _loggerService.Error($"Unable to find user with id {userId}");
+            return null;
+        }
         Board result = _boardService.RemoveMember(board, user);
         return result;
 
+    }
+    public Board UpdateBoardName(string boardId,string name){
+        Board board=_boardService.GetBoard(boardId);
+        if(board is null){
+            _loggerService.Error($"Board with id {boardId} not present.");
+        }
+        board.Name = name;
+        board=_boardService.UpdateBoard(board);
+        return board;
+    }
+    public Board UpdateBoardPrivary(string boardId,Privacy privacy){
+        Board board=_boardService.GetBoard(boardId);
+        if(board is null){
+            _loggerService.Error($"Board with id {boardId} is not found");
+            return null;
+        }
+        else{
+            board.Privacy=privacy;
+            board=_boardService.UpdateBoard(board);
+            return board;
+        }
+    }
+    public string ShowBoard(string boardId){
+        Board board=_boardService.GetBoard(boardId);
+        if(board is null){
+            _loggerService.Error($"Board with id {boardId} is not present.");
+            return null;
+        }
+        return board.ToString();
+        
     }
 
     public User AddUser(string name, string email)
@@ -81,9 +138,7 @@ public class Trello
     }
 
 
-
-
-    public BoardList AddList(string boardId, string Name)
+    public BoardList CreateList(string boardId, string Name)
     {
         Board board = _boardService.GetBoard(boardId);
         if (board is null)
@@ -94,7 +149,104 @@ public class Trello
         BoardList boardList = new BoardList();
         boardList.Name = Name;
         boardList.Board = board;
+    
         BoardList result = _listService.AddList(boardList);
         return result;
+    }
+    public BoardList RemoveList(string listId){
+        BoardList list=_listService.GetBoardList(listId);
+        if(list is null){
+            _loggerService.Error($" List with id {listId} is not present.");
+        }
+        return _listService.RemoveList(list);
+    }
+    public BoardList UpdateListName(string listId,string name){
+        BoardList list=_listService.GetBoardList(listId);
+        if(list is null){
+            _loggerService.Error($" board list with id {listId} is not present.");
+            return null;
+        }
+        list.Name=name;
+        return _listService.UpdateBoardList(list);
+    }
+    public string ShowBoardList(string listId){
+        BoardList boardList=_listService.GetBoardList(listId);
+        if(boardList is null){
+            _loggerService.Error($"Board list with id {listId} is not present.");
+            return null;
+        }
+        return boardList.ToString();
+    }
+    
+    
+    public Card CreateCard(string listId,string name, string description,User assigneduser){
+        BoardList boardList=_listService.GetBoardList(listId);
+        if(boardList is null){
+            _loggerService.Error($"Given board list id {listId} is not present.");
+            return null;
+
+        }
+        Card card=new Card();
+        card.Id=_idGeneratorService.GenerateId();
+        card.Name=name;
+        card.BoardList=boardList;
+        card.AssignedUser=assigneduser;
+        card.Description=description;
+        card=_cardService.AddCard(card);
+        return card;
+
+    }
+    public Card DeleteCard(string cardId){
+        Card card=_cardService.GetCard(cardId);
+        if(card is null){
+            _loggerService.Error($"Card with given id {cardId} not present.");
+            return null;
+        }
+        return card;
+    }
+    public Card UpdateCardName(string cardId,string name){
+        Card card=_cardService.GetCard(cardId);
+        if(card is null){
+            _loggerService.Error($"card with {cardId} is not present");
+            return null;
+        }
+        card.Name=name;
+        card=_cardService.UpdateCard(card);
+        return card;
+    }
+    public Card UpdateCardDescription(string cardId,string description){
+        Card card=_cardService.GetCard(cardId);
+        if(card is null){
+            _loggerService.Error($"card with {cardId} is not present");
+            return null;
+        }
+        card.Description=description;
+        card=_cardService.UpdateCard(card);
+        return card;
+    }
+    public Card UpdateAssignedUser(string cardId,string userId){
+        Card card=_cardService.GetCard(cardId);
+        User user=_userService.GetUser(userId);
+        if(user is null){
+            _loggerService.Error($"user with is {userId} not found");
+            return null;
+        }
+        if(card is null){
+            _loggerService.Error($"card with {cardId} is not present");
+            return null;
+        }
+        card.AssignedUser=user;
+        card=_cardService.UpdateCard(card);
+        return card;
+    }
+    public string ShowCard(string id){
+        Card? card=_cardService.GetCard(id);
+        if(card is null){
+            _loggerService.Warning($"Card with id {id} is not present.");
+            return null;
+        }
+        else{
+            return card.ToString();
+        }
     }
 }
